refactor: move code that calculates vis config to separate module
